+++ chalicelib/blueprint.py
@@ -1,16 +1,16 @@
 import openai
 from . import pvsecret
 import os
-from chalicelib.constants import API_VERSION
+from chalicelib.version import API_VERSION
 
 blueprint_api_version = API_VERSION  # API version is global for now, not service specific
+print("blueprint_api_version: ", blueprint_api_version)
 
 secret_json = pvsecret.get_secrets()
 
 # TEMP - put this back to the polyverse key once gpt-4 access is approved there
 openai_key = secret_json["openai-personal"]
 openai.api_key = openai_key
-print("openai key ", openai_key)
 
 # Define the directory where prompt files are stored
 PROMPT_DIR = "chalicelib/prompts"
@@ -24,7 +24,6 @@ ROLE_SYSTEM_FILENAME = "blueprint-role-system.prompt"
 # Load the prompt files and replace the placeholders with the actual values
 def load_prompts():
     promptdir = os.path.join(os.path.abspath(os.path.curdir), PROMPT_DIR)
-    print("promptdir: " + promptdir)
 
     # Load the prompt file for seed
     with open(os.path.join(promptdir, SEED_PROMPT_FILENAME), 'r') as f:
@@ -60,8 +59,6 @@ def blueprint_code(json_data):
             prompt = blueprint_seed_prompt.format(code=code, prior_blueprint=prior_blueprint)
     else:
         prompt = blueprint_seed_prompt.format(code=code)
-
-    print("calling openai with prompt: " + prompt + "\n\n")
 
     response = openai.ChatCompletion.create(
         model="gpt-4",