+++ chalicelib/constants.py
@@ -1,1 +1,0 @@
-API_VERSION = "0.4.2"  # API version is global for now, not service specific